Merge

+++ beacon/include/marlin/beacon/ClusterDiscoverer.hpp
@@ -18,9 +18,9 @@ static std::array<uint8_t, size> hextoint(std::string s){
 	//hex strings starting with 0x//
 	assert(size >= (s.length()-2)/2);
 	std::array<uint8_t, size> res;
-	for(int i=2;i<s.length();i+=2){
+	for(size_t i = 2; i < s.length(); i += 2){
 		char tmp[2] = {s[i], s[i+1]};
-		res[(i-2)/2] = (uint8_t)strtol(tmp, NULL, 16); 
+		res[(i-2)/2] = (uint8_t)strtol(tmp, NULL, 16);
 	}
 	return res;
 }
